feat: 신고 기능 구현 및 CSS 변수 리팩토링
문장 신고 기능과 신고 내역 페이지 구현, CSS 변수 시스템 도입

주요 기능:

1. 문장 신고 기능
   - QuoteMoreMenu: Quote 컴포넌트에 더보기 버튼(⋮) 및 드롭다운 메뉴 추가
   - ReportPopup: 신고 사유 선택(수정요청/삭제요청), 상세설명(선택) 입력
   - 로그인 상태 확인 후 미로그인 시 LoginRequiredPopup 표시
   - 한글 IME 조합 중 글자수 초과 방지

2. 신고 내역 페이지 (/quote/report)
   - MyReports: 무한 스크롤, 상태별 필터링(전체/대기중/처리완료)
   - ReportCard: 신고 사유/상태 뱃지, 문장 미리보기, 삭제 버튼
   - ReportFilters: 필터 버튼 그룹
   - ProfileDropdown에 "신고 내역" 메뉴 추가

3. 공통 컴포넌트
   - Badge: 재사용 가능한 뱃지 컴포넌트
   - LoginRequiredPopup: 커스텀 메시지 지원 추가
   - formatDate: 날짜 포맷 유틸리티

4. CSS 변수 리팩토링
   - index.css: primary-light, primary-bg, bg-tertiary, danger 색상 추가
   - 다크모드 변수 동기화
   - ReportPopup, ReportCard, QuoteMoreMenu 등 CSS 변수 적용

5. QuoteContext 개선
   - currentQuote 변수 추가 (현재 표시 중인 문장 접근용)
