+++ tp-react/src/components/LoginRequiredPopup/LoginRequiredPopup.css
@@ -1,0 +1,85 @@
+.login-required-overlay {
+    position: fixed;
+    top: 0;
+    left: 0;
+    width: 100%;
+    height: 100%;
+    background-color: rgba(0, 0, 0, 0.5);
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    z-index: 2500;
+}
+
+.login-required-popup {
+    background-color: var(--color-bg);
+    border-radius: 0.75rem;
+    padding: 1.5rem;
+    width: 90%;
+    max-width: 400px;
+    border: 1px solid var(--color-border);
+}
+
+.login-required-popup.dark {
+    background-color: #0f0f0f;
+    border-color: #27272a;
+}
+
+.login-required-message {
+    font-size: 1rem;
+    color: var(--color-text);
+    text-align: center;
+    margin: 0 0 1.5rem 0;
+}
+
+.login-required-popup.dark .login-required-message {
+    color: #e5e7eb;
+}
+
+.login-required-actions {
+    display: flex;
+    gap: 0.75rem;
+}
+
+.login-required-cancel-btn {
+    flex: 1;
+    padding: 0.75rem;
+    border: 1px solid var(--color-border);
+    border-radius: 0.5rem;
+    background-color: transparent;
+    color: var(--color-text-muted);
+    font-size: 1rem;
+    cursor: pointer;
+    transition: all 0.2s;
+}
+
+.login-required-cancel-btn:hover {
+    background-color: var(--color-bg-secondary);
+    color: var(--color-text);
+}
+
+.login-required-cancel-btn.dark {
+    border-color: #27272a;
+    color: #a1a1aa;
+}
+
+.login-required-cancel-btn.dark:hover {
+    background-color: #18181b;
+    color: #e5e7eb;
+}
+
+.login-required-login-btn {
+    flex: 1;
+    padding: 0.75rem;
+    border: none;
+    border-radius: 0.5rem;
+    background-color: var(--color-primary);
+    color: var(--color-white);
+    font-size: 1rem;
+    cursor: pointer;
+    transition: background-color 0.2s;
+}
+
+.login-required-login-btn:hover {
+    background-color: var(--color-primary-dark);
+}
