+++ tp-react/src/components/LoginRequiredPopup/LoginRequiredPopup.jsx
@@ -1,0 +1,43 @@
+import {useTheme} from '../../Context/ThemeContext';
+import {useAuth} from '../../Context/AuthContext';
+import './LoginRequiredPopup.css';
+
+const LoginRequiredPopup = ({message, onClose}) => {
+    const {isDark} = useTheme();
+    const {triggerLogin} = useAuth();
+
+    const handleOverlayClick = (e) => {
+        if (e.target === e.currentTarget) {
+            onClose();
+        }
+    };
+
+    const handleLogin = () => {
+        onClose();
+        triggerLogin();
+    };
+
+    return (
+        <div className={`login-required-overlay ${isDark ? 'dark' : ''}`} onClick={handleOverlayClick}>
+            <div className={`login-required-popup ${isDark ? 'dark' : ''}`}>
+                <p className="login-required-message">{message}</p>
+                <div className="login-required-actions">
+                    <button
+                        className={`login-required-cancel-btn ${isDark ? 'dark' : ''}`}
+                        onClick={onClose}
+                    >
+                        취소
+                    </button>
+                    <button
+                        className="login-required-login-btn"
+                        onClick={handleLogin}
+                    >
+                        로그인
+                    </button>
+                </div>
+            </div>
+        </div>
+    );
+};
+
+export default LoginRequiredPopup;
