+++ tp-react/src/components/Quote/Quote.css
@@ -23,7 +23,8 @@
 
 .author-container {
     display: flex;
-    flex-direction: row-reverse;
+    align-items: center;
+    gap: 0.25rem;
 }
 
 .sentence-input-wrapper {
