+++ tp-react/src/components/ReportPopup/ReportPopup.css
@@ -1,0 +1,285 @@
+.report-popup-overlay {
+    position: fixed;
+    top: 0;
+    left: 0;
+    width: 100%;
+    height: 100%;
+    background-color: rgba(0, 0, 0, 0.5);
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    z-index: 2000;
+}
+
+.report-popup {
+    background-color: var(--color-white);
+    border-radius: 1rem;
+    padding: 1.5rem;
+    max-width: 480px;
+    width: 90%;
+    max-height: 90vh;
+    overflow-y: auto;
+    box-shadow: 0 25px 50px -12px rgba(0, 0, 0, 0.25);
+}
+
+.report-popup.dark {
+    background-color: var(--color-bg-secondary);
+}
+
+.report-popup-header {
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    margin-bottom: 1.25rem;
+}
+
+.report-popup-title {
+    font-size: 1.25rem;
+    font-weight: 600;
+    color: var(--color-text);
+}
+
+.report-popup-close-btn {
+    background: none;
+    border: none;
+    color: var(--color-text-placeholder);
+    font-size: 1.25rem;
+    cursor: pointer;
+    padding: 0.25rem;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    transition: color 0.2s;
+}
+
+.report-popup-close-btn:hover {
+    color: var(--color-text-muted);
+}
+
+/* 신고 대상 문장 미리보기 */
+.report-quote-preview {
+    background-color: var(--color-bg-secondary);
+    border-radius: 0.5rem;
+    padding: 1rem;
+    margin-bottom: 1.25rem;
+}
+
+.report-quote-preview.dark {
+    background-color: var(--color-bg-tertiary);
+}
+
+.report-quote-label {
+    display: block;
+    font-size: 0.75rem;
+    color: var(--color-text-placeholder);
+    margin-bottom: 0.5rem;
+}
+
+.report-quote-sentence {
+    font-size: 0.95rem;
+    color: var(--color-text);
+    line-height: 1.5;
+    margin: 0;
+}
+
+.report-quote-author {
+    display: block;
+    font-size: 0.85rem;
+    color: var(--color-text-muted);
+    margin-top: 0.5rem;
+}
+
+/* 신고 사유 선택 */
+.report-reason-section {
+    margin-bottom: 1.25rem;
+}
+
+.report-section-label {
+    display: block;
+    font-size: 0.9rem;
+    font-weight: 500;
+    color: var(--color-text);
+    margin-bottom: 0.75rem;
+}
+
+.report-optional {
+    font-weight: 400;
+    color: var(--color-text-placeholder);
+}
+
+.report-reason-options {
+    display: flex;
+    flex-direction: column;
+    gap: 0.5rem;
+}
+
+.report-reason-option {
+    display: flex;
+    align-items: flex-start;
+    gap: 0.75rem;
+    padding: 0.75rem;
+    border: 1px solid var(--color-border);
+    border-radius: 0.5rem;
+    cursor: pointer;
+    transition: all 0.2s;
+}
+
+.report-reason-option:hover {
+    border-color: var(--color-primary);
+}
+
+.report-reason-option:has(input:checked) {
+    border-color: var(--color-primary);
+    background-color: var(--color-primary-bg);
+}
+
+.report-reason-option.dark:has(input:checked) {
+    background-color: var(--color-primary-bg-dark);
+}
+
+.report-reason-option input[type="radio"] {
+    display: none;
+}
+
+.report-reason-radio {
+    width: 18px;
+    height: 18px;
+    border: 2px solid var(--color-border);
+    border-radius: 50%;
+    flex-shrink: 0;
+    position: relative;
+    margin-top: 0.125rem;
+}
+
+.report-reason-option:has(input:checked) .report-reason-radio {
+    border-color: var(--color-primary);
+}
+
+.report-reason-option:has(input:checked) .report-reason-radio::after {
+    content: '';
+    position: absolute;
+    top: 50%;
+    left: 50%;
+    transform: translate(-50%, -50%);
+    width: 10px;
+    height: 10px;
+    background-color: var(--color-primary);
+    border-radius: 50%;
+}
+
+.report-reason-content {
+    display: flex;
+    flex-direction: column;
+    gap: 0.25rem;
+}
+
+.report-reason-text {
+    font-size: 0.95rem;
+    font-weight: 500;
+    color: var(--color-text);
+    display: block;
+}
+
+.report-reason-desc {
+    font-size: 0.8rem;
+    color: var(--color-text-muted);
+    display: block;
+}
+
+/* 상세 설명 */
+.report-detail-section {
+    margin-bottom: 1.25rem;
+    position: relative;
+}
+
+.report-detail-input {
+    width: 100%;
+    padding: 0.75rem;
+    border: 1px solid var(--color-border);
+    border-radius: 0.5rem;
+    font-size: 0.95rem;
+    font-family: inherit;
+    resize: none;
+    outline: none;
+    transition: border-color 0.2s;
+    background-color: var(--color-white);
+    color: var(--color-text);
+}
+
+.report-detail-input:focus {
+    border-color: var(--color-primary);
+}
+
+.report-detail-input::placeholder {
+    color: var(--color-text-placeholder);
+}
+
+.report-detail-input.dark {
+    background-color: var(--color-bg-tertiary);
+    border-color: var(--color-border);
+}
+
+.report-detail-input.dark:focus {
+    border-color: var(--color-primary-light);
+}
+
+.report-detail-count {
+    position: absolute;
+    bottom: 0.5rem;
+    right: 0.75rem;
+    font-size: 0.75rem;
+    color: var(--color-text-placeholder);
+}
+
+/* 버튼 */
+.report-popup-actions {
+    display: flex;
+    gap: 0.75rem;
+}
+
+.report-cancel-btn {
+    flex: 1;
+    padding: 0.75rem;
+    border: 1px solid var(--color-border);
+    border-radius: 0.5rem;
+    background-color: transparent;
+    color: var(--color-text-muted);
+    font-size: 0.95rem;
+    cursor: pointer;
+    transition: all 0.2s;
+}
+
+.report-cancel-btn:hover:not(:disabled) {
+    background-color: var(--color-bg-secondary);
+    color: var(--color-text);
+}
+
+.report-cancel-btn:disabled {
+    opacity: 0.5;
+    cursor: not-allowed;
+}
+
+.report-submit-btn {
+    flex: 1;
+    padding: 0.75rem;
+    border: none;
+    border-radius: 0.5rem;
+    background-color: var(--color-danger);
+    color: var(--color-white);
+    font-size: 0.95rem;
+    cursor: pointer;
+    transition: background-color 0.2s;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    gap: 0.5rem;
+}
+
+.report-submit-btn:hover:not(:disabled) {
+    background-color: var(--color-danger-dark);
+}
+
+.report-submit-btn:disabled {
+    background-color: var(--color-danger-light);
+    cursor: not-allowed;
+}
