+++ tp-react/src/components/ReportPopup/ReportReasonOption.jsx
@@ -1,0 +1,24 @@
+import {useTheme} from '../../Context/ThemeContext';
+
+const ReportReasonOption = ({value, checked, onChange, title, description}) => {
+    const {isDark} = useTheme();
+
+    return (
+        <label className={`report-reason-option ${isDark ? 'dark' : ''}`}>
+            <input
+                type="radio"
+                name="reportReason"
+                value={value}
+                checked={checked}
+                onChange={onChange}
+            />
+            <span className="report-reason-radio"></span>
+            <div className="report-reason-content">
+                <span className="report-reason-text">{title}</span>
+                <span className={`report-reason-desc ${isDark ? 'dark' : ''}`}>{description}</span>
+            </div>
+        </label>
+    );
+};
+
+export default ReportReasonOption;
