+++ tp-react/src/index.css
@@ -5,6 +5,9 @@
     /* Primary (Purple) */
     --color-primary: #6d28d9;
     --color-primary-dark: #5b21b6;
+    --color-primary-light: #8b5cf6;
+    --color-primary-bg: #f5f3ff;
+    --color-primary-bg-dark: rgba(139, 92, 246, 0.1);
 
     /* Text */
     --color-text: #333;
@@ -14,6 +17,7 @@
     /* Background */
     --color-bg: #fafafa;
     --color-bg-secondary: #f4f4f5;
+    --color-bg-tertiary: #e4e4e7;
     --color-popup-bg: #ffffff;
 
     /* Border & Toggle */
@@ -28,6 +32,11 @@
     --color-success: #16a34a;
     --color-correct: #333;
     --color-white: #ffffff;
+
+    /* Danger (for destructive actions) */
+    --color-danger: #ef4444;
+    --color-danger-dark: #dc2626;
+    --color-danger-light: #fca5a5;
 
     /* Badge - Public (Blue) */
     --color-badge-public-bg: #dbeafe;
@@ -44,13 +53,28 @@
     /* Badge - Active (Green) */
     --color-badge-active-bg: #dcfce7;
     --color-badge-active-text: #16a34a;
+
+    /* Badge - Processed (Green) */
+    --color-badge-processed-bg: #d1fae5;
+    --color-badge-processed-text: #059669;
+
+    /* Badge - Reason Modify (Blue) */
+    --color-badge-modify-bg: #dbeafe;
+    --color-badge-modify-text: #1d4ed8;
+
+    /* Badge - Reason Delete (Red) */
+    --color-badge-delete-bg: #fee2e2;
+    --color-badge-delete-text: #dc2626;
 }
 
 /* Dark Mode */
 .dark {
-    /* Primary (Purple) - same as light */
+    /* Primary (Purple) */
     --color-primary: #6d28d9;
     --color-primary-dark: #5b21b6;
+    --color-primary-light: #8b5cf6;
+    --color-primary-bg: rgba(139, 92, 246, 0.15);
+    --color-primary-bg-dark: rgba(139, 92, 246, 0.1);
 
     /* Text */
     --color-text: #e5e7eb;
@@ -60,10 +84,11 @@
     /* Background */
     --color-bg: #0f0f0f;
     --color-bg-secondary: #18181b;
+    --color-bg-tertiary: #27272a;
     --color-popup-bg: #1f1f1f;
 
     /* Border & Toggle */
-    --color-border: #27272a;
+    --color-border: #3f3f46;
     --color-toggle-bg: #3f3f46;
 
     /* Semantic */
@@ -73,6 +98,11 @@
     --color-error-border: #f87171;
     --color-success: #22c55e;
     --color-correct: #ffffff;
+
+    /* Danger (for destructive actions) */
+    --color-danger: #ef4444;
+    --color-danger-dark: #dc2626;
+    --color-danger-light: #7f1d1d;
 
     /* Badge - Public (Blue) */
     --color-badge-public-bg: #1e3a5f;
@@ -89,6 +119,18 @@
     /* Badge - Active (Green) */
     --color-badge-active-bg: #1a3d2e;
     --color-badge-active-text: #86efac;
+
+    /* Badge - Processed (Green) */
+    --color-badge-processed-bg: #064e3b;
+    --color-badge-processed-text: #6ee7b7;
+
+    /* Badge - Reason Modify (Blue) */
+    --color-badge-modify-bg: #1e3a5f;
+    --color-badge-modify-text: #93c5fd;
+
+    /* Badge - Reason Delete (Red) */
+    --color-badge-delete-bg: #450a0a;
+    --color-badge-delete-text: #fca5a5;
 }
 
 /* ===========================================
