+++ tp-react/src/pages/MyReports/MyReports.css
@@ -1,0 +1,118 @@
+.my-reports-container {
+    width: 80%;
+    max-width: 1000px;
+    min-width: 600px;
+    margin: 0 auto;
+    padding: 2rem;
+}
+
+.my-reports-header {
+    margin-bottom: 1.5rem;
+}
+
+.my-reports-title {
+    font-size: 1.5rem;
+    font-weight: 600;
+    color: var(--color-primary);
+    margin: 0;
+}
+
+/* 로그인 필요 */
+.my-reports-login-required {
+    text-align: center;
+    padding: 3rem;
+    color: var(--color-text-muted);
+}
+
+.my-reports-login-required button {
+    margin-top: 1rem;
+    padding: 0.75rem 1.5rem;
+    border: none;
+    border-radius: 0.5rem;
+    background-color: var(--color-primary);
+    color: var(--color-white);
+    font-size: 1rem;
+    cursor: pointer;
+    transition: background-color 0.2s;
+}
+
+.my-reports-login-required button:hover {
+    background-color: var(--color-primary-dark);
+}
+
+/* 신고 목록 */
+.my-reports-list {
+    display: flex;
+    flex-direction: column;
+    gap: 1rem;
+}
+
+/* 로딩 */
+.my-reports-loading {
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    padding: 1rem;
+}
+
+.my-reports-spinner {
+    width: 1.25rem;
+    height: 1.25rem;
+    border: 2px solid var(--color-border);
+    border-top-color: var(--color-primary);
+    border-radius: 50%;
+    animation: my-reports-spin 0.8s linear infinite;
+}
+
+@keyframes my-reports-spin {
+    0% {
+        transform: rotate(0deg);
+    }
+    100% {
+        transform: rotate(360deg);
+    }
+}
+
+/* 빈 상태 */
+.my-reports-empty {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+    padding: 3rem;
+    color: var(--color-text-placeholder);
+}
+
+.my-reports-empty svg {
+    font-size: 3rem;
+    margin-bottom: 1rem;
+}
+
+.my-reports-empty p {
+    font-size: 1rem;
+    margin: 0;
+}
+
+/* 하단 버튼 */
+.my-reports-actions {
+    display: flex;
+    justify-content: center;
+    gap: 0.75rem;
+    padding-top: 1.5rem;
+}
+
+.my-reports-back-btn {
+    padding: 0.75rem 2rem;
+    border: 1px solid var(--color-border);
+    border-radius: 0.5rem;
+    background-color: transparent;
+    color: var(--color-text-muted);
+    font-size: 1rem;
+    cursor: pointer;
+    transition: all 0.2s;
+}
+
+.my-reports-back-btn:hover {
+    background-color: var(--color-bg-secondary);
+    color: var(--color-text);
+}
