+++ tp-react/src/pages/MyReports/components/ReportCard.css
@@ -1,0 +1,126 @@
+.report-card {
+    background-color: var(--color-white);
+    border: 1px solid var(--color-border);
+    border-radius: 0.75rem;
+    padding: 1rem;
+}
+
+.report-card.dark {
+    background-color: var(--color-bg-secondary);
+}
+
+.report-card-header {
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    margin-bottom: 0.75rem;
+}
+
+.report-badges {
+    display: flex;
+    gap: 0.5rem;
+}
+
+.report-badge {
+    padding: 0.25rem 0.5rem;
+    border-radius: 0.25rem;
+    font-size: 0.75rem;
+    font-weight: 500;
+}
+
+.report-badge.reason-modify {
+    background-color: var(--color-badge-modify-bg);
+    color: var(--color-badge-modify-text);
+}
+
+.report-badge.reason-delete {
+    background-color: var(--color-badge-delete-bg);
+    color: var(--color-badge-delete-text);
+}
+
+.report-badge.status-pending {
+    background-color: var(--color-badge-pending-bg);
+    color: var(--color-badge-pending-text);
+}
+
+.report-badge.status-processed {
+    background-color: var(--color-badge-processed-bg);
+    color: var(--color-badge-processed-text);
+}
+
+.report-delete-btn {
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    width: 2rem;
+    height: 2rem;
+    border: none;
+    border-radius: 0.375rem;
+    background-color: transparent;
+    color: var(--color-text-placeholder);
+    cursor: pointer;
+    transition: all 0.2s;
+}
+
+.report-delete-btn:hover {
+    background-color: var(--color-badge-delete-bg);
+    color: var(--color-badge-delete-text);
+}
+
+/* 신고 대상 문장 */
+.report-quote {
+    background-color: var(--color-bg-secondary);
+    border-radius: 0.5rem;
+    padding: 0.75rem;
+    margin-bottom: 0.75rem;
+}
+
+.report-quote.dark {
+    background-color: var(--color-bg-tertiary);
+}
+
+.report-quote.quote-deleted {
+    opacity: 0.6;
+}
+
+.report-sentence {
+    font-size: 0.9rem;
+    color: var(--color-text);
+    margin: 0 0 0.25rem 0;
+    line-height: 1.5;
+}
+
+.report-author {
+    font-size: 0.8rem;
+    color: var(--color-text-muted);
+}
+
+/* 신고 내용 */
+.report-detail {
+    margin-bottom: 0.75rem;
+}
+
+.report-detail-label {
+    display: block;
+    font-size: 0.75rem;
+    color: var(--color-text-placeholder);
+    margin-bottom: 0.25rem;
+}
+
+.report-detail-text {
+    font-size: 0.875rem;
+    color: var(--color-text-muted);
+    margin: 0;
+    line-height: 1.5;
+}
+
+/* 푸터 */
+.report-card-footer {
+    display: flex;
+    justify-content: flex-end;
+}
+
+.report-date {
+    font-size: 0.75rem;
+    color: var(--color-text-placeholder);
+}
