+++ tp-react/src/pages/MyReports/components/ReportFilters.css
@@ -1,0 +1,37 @@
+.report-filters {
+    display: flex;
+    gap: 1rem;
+    margin-bottom: 1.5rem;
+}
+
+.report-filter-group {
+    display: flex;
+    gap: 0.5rem;
+}
+
+.report-filter-btn {
+    padding: 0.5rem 1rem;
+    border: 1px solid var(--color-border);
+    border-radius: 0.375rem;
+    background-color: transparent;
+    color: var(--color-text-muted);
+    font-size: 0.875rem;
+    cursor: pointer;
+    transition: all 0.2s;
+}
+
+.report-filter-btn:hover {
+    border-color: var(--color-text-muted);
+    color: var(--color-text);
+}
+
+.report-filter-btn.active {
+    border-color: var(--color-primary);
+    background-color: var(--color-primary);
+    color: var(--color-white);
+}
+
+.report-filter-btn.dark.active {
+    border-color: var(--color-primary-light);
+    background-color: var(--color-primary-light);
+}
