+++ tp-react/src/utils/formatDate.js
@@ -1,0 +1,12 @@
+/**
+ * 날짜를 YYYY.MM.DD 형식으로 포맷
+ * @param {string|Date} dateString - 날짜 문자열 또는 Date 객체
+ * @returns {string} 포맷된 날짜 문자열
+ */
+export const formatDate = (dateString) => {
+    const date = new Date(dateString);
+    const year = date.getFullYear();
+    const month = String(date.getMonth() + 1).padStart(2, '0');
+    const day = String(date.getDate()).padStart(2, '0');
+    return `${year}.${month}.${day}`;
+};
